[FLINK-14149][coordination] Integrate ZooKeeperLeaderElectionServiceNG
