+++ flink-runtime/src/main/java/org/apache/flink/runtime/highavailability/HighAvailabilityServicesUtils.java
@@ -147,7 +147,7 @@ public class HighAvailabilityServicesUtils {
 				return new StandaloneClientHAServices(webMonitorAddress);
 			case ZOOKEEPER:
 				final CuratorFramework client = ZooKeeperUtils.startCuratorFramework(configuration);
-				return new ZooKeeperClientHAServices(client, configuration);
+				return new ZooKeeperClientHAServices(client);
 			case FACTORY_CLASS:
 				return createCustomClientHAServices(configuration);
 			default:
