+++ flink-runtime/src/main/java/org/apache/flink/runtime/highavailability/zookeeper/ZooKeeperClientHAServices.java
@@ -18,35 +18,28 @@
 
 package org.apache.flink.runtime.highavailability.zookeeper;
 
-import org.apache.flink.configuration.Configuration;
 import org.apache.flink.runtime.highavailability.ClientHighAvailabilityServices;
 import org.apache.flink.runtime.leaderretrieval.LeaderRetrievalService;
-import org.apache.flink.runtime.util.ZooKeeperUtils;
+import org.apache.flink.runtime.leaderretrieval.ZooKeeperLeaderRetrievalService;
 
 import org.apache.curator.framework.CuratorFramework;
-
-import javax.annotation.Nonnull;
 
 /**
  * ZooKeeper based implementation for {@link ClientHighAvailabilityServices}.
  */
 public class ZooKeeperClientHAServices implements ClientHighAvailabilityServices {
 
-	private static final String REST_SERVER_LEADER_PATH = "/rest_server_lock";
+	private static final String REST_SERVER_LEADER_PATH = "/rest-server";
 
 	private final CuratorFramework client;
-	private final Configuration configuration;
 
-	public ZooKeeperClientHAServices(
-		@Nonnull CuratorFramework client,
-		@Nonnull Configuration configuration) {
+	public ZooKeeperClientHAServices(CuratorFramework client) {
 		this.client = client;
-		this.configuration = configuration;
 	}
 
 	@Override
 	public LeaderRetrievalService getClusterRestEndpointLeaderRetriever() {
-		return ZooKeeperUtils.createLeaderRetrievalService(client, configuration, REST_SERVER_LEADER_PATH);
+		return new ZooKeeperLeaderRetrievalService(client, REST_SERVER_LEADER_PATH);
 	}
 
 	@Override
